misc fixes here and there
