+++ src/rdf4cpp/rdf/datatypes/registry/DatatypeID.hpp
@@ -7,6 +7,8 @@
 
 #include <cstdint>
 #include <functional>
+#include <string>
+#include <string_view>
 #include <type_traits>
 #include <variant>
