+++ src/rdf4cpp/rdf/datatypes/registry/DatatypeRegistry.hpp
@@ -31,7 +31,6 @@ public:
     using factory_fptr_t = std::any (*)(std::string_view);
     using to_string_fptr_t = std::string (*)(const std::any &) noexcept;
     using ebv_fptr_t = bool (*)(std::any const &) noexcept;
-    using can_inline_fptr_t = bool (*)(std::any const &) noexcept;
     using try_into_inlined_fptr_t = std::optional<uint64_t> (*)(std::any const &) noexcept;
     using from_inlined_fptr_t = std::any (*)(uint64_t) noexcept;
 
@@ -155,16 +154,7 @@ public:
 private:
     using registered_datatypes_t = std::vector<DatatypeEntry>;
 
-    inline static registered_datatypes_t &get_mutable() noexcept {
-        static registered_datatypes_t registry_ = []() {
-            registered_datatypes_t r;
-            r.resize(dynamic_datatype_offset, DatatypeEntry::placeholder()); // placeholders for fixed id datatypes
-
-            return r;
-        }();
-
-        return registry_;
-    }
+    static registered_datatypes_t &get_mutable() noexcept;
 
     /**
      * Tries to find the datatype corresponding to datatype_id
@@ -176,30 +166,7 @@ private:
      */
     template<typename Map>
         requires std::invocable<Map, DatatypeEntry const &>
-    static std::optional<std::invoke_result_t<Map, DatatypeEntry const &>> find_map_entry(DatatypeIDView const datatype_id, Map f) noexcept(std::is_nothrow_invocable_v<Map, DatatypeEntry const &>) {
-        using ret_type = std::optional<std::invoke_result_t<Map, DatatypeEntry const &>>;
-        auto const &registry = registered_datatypes();
-
-        return visit(DatatypeIDVisitor{
-                             [&registry, f](LiteralType const fixed_id) -> ret_type {
-                                 auto const id_as_index = static_cast<size_t>(fixed_id.to_underlying()) - 1;  // ids from 1 to n stored in places 0 to n-1
-                                 assert(id_as_index < dynamic_datatype_offset);
-
-                                 return f(registry[id_as_index]);
-                             },
-                             [&registry, f](std::string_view const other_iri) -> ret_type {
-                                 auto found = std::lower_bound(registry.begin() + dynamic_datatype_offset, registry.end(),
-                                                               DatatypeEntry::for_search(other_iri),
-                                                               [](const auto &left, const auto &right) { return left.datatype_iri < right.datatype_iri; });
-
-                                 if (found != registry.end() and found->datatype_iri == other_iri) {
-                                     return f(*found);
-                                 } else {
-                                     return std::nullopt;
-                                 }
-                             }},
-                     datatype_id);
-    }
+    static std::optional<std::invoke_result_t<Map, DatatypeEntry const &>> find_map_entry(DatatypeIDView datatype_id, Map f) noexcept(std::is_nothrow_invocable_v<Map, DatatypeEntry const &>);
 
     /**
      * Creates NumericOps based on a NumericLiteralDatatype
@@ -211,14 +178,7 @@ private:
     template<datatypes::InlineableLiteralDatatype LiteralDatatype_t>
     static InliningOps make_inlining_ops() noexcept;
 
-    inline static void add_fixed(DatatypeEntry entry_to_add, LiteralType type_id) noexcept {
-        auto const id_as_index = static_cast<size_t>(type_id.to_underlying()) - 1; // ids from 1 to n stored in places 0 to n-1
-        assert(id_as_index < dynamic_datatype_offset);
-
-        auto &slot = DatatypeRegistry::get_mutable()[id_as_index];
-        assert(slot.datatype_iri.empty()); // is placeholder
-        slot = std::move(entry_to_add);
-    }
+    static void add_fixed(DatatypeEntry entry_to_add, LiteralType type_id) noexcept;
 
 public:
     /**
@@ -226,77 +186,47 @@ public:
      * @tparam datatype_info type that is registered.
      */
     template<datatypes::LiteralDatatype LiteralDatatype_t>
-    inline static void add() noexcept;
+    static void add() noexcept;
 
     /**
      * Register an datatype manually
      */
-    inline static void add(DatatypeEntry entry_to_add) noexcept {
-        auto &registry = DatatypeRegistry::get_mutable();
-
-        auto found = std::find_if(registry.begin() + dynamic_datatype_offset, registry.end(), [&](const auto &entry) { return entry.datatype_iri == entry_to_add.datatype_iri; });
-        if (found == registry.end()) {
-            registry.push_back(std::move(entry_to_add));
-
-            std::sort(registry.begin() + dynamic_datatype_offset, registry.end(),
-                      [](const auto &left, const auto &right) { return left.datatype_iri < right.datatype_iri; });
-        } else {
-            *found = std::move(entry_to_add);
-        }
-    }
+    static void add(DatatypeEntry entry_to_add) noexcept;
 
     /**
      * Retrieve all registered datatypes.
      * @return vector of DatatypeEntries
      */
-    inline static const registered_datatypes_t &registered_datatypes() noexcept {
-        return DatatypeRegistry::get_mutable();
-    }
-
-    inline static std::optional<std::string_view> get_iri(DatatypeIDView const datatype_id) noexcept {
-        return find_map_entry(datatype_id, [](auto const &entry) noexcept -> std::string_view {
-            return entry.datatype_iri;
-        });
-    }
+    [[nodiscard]] static registered_datatypes_t const &registered_datatypes() noexcept;
 
     /**
      * Get the database entry for a datatype_id.
-     * @param datatype_id datatype IRI string
+     * @param datatype_id datatype id of the datatype
      * @return if available database entry else nullopt
      */
-    inline static DatatypeEntry const *get_entry(DatatypeIDView const datatype_id) noexcept {
-        auto const res = find_map_entry(datatype_id, [](auto const &entry) noexcept {
-            return &entry;
-        });
+    [[nodiscard]] static DatatypeEntry const *get_entry(DatatypeIDView datatype_id) noexcept;
 
-        return res.has_value() ? *res : nullptr;
-    }
+    /**
+     * Get the IRI for a datatype_id
+     *
+     * @param datatype_id datatype id of the datatype
+     * @return if the type exists then its IRI else nullopt
+     */
+    [[nodiscard]] static std::optional<std::string_view> get_iri(DatatypeIDView datatype_id) noexcept;
 
     /**
      * Get a factory_fptr_t for a datatype. The factory_fptr_t can be used like `std::any type_instance = factory_fptr("types string representation")`.
      * @param datatype_id datatype id for the corresponding datatype
      * @return function pointer or nullptr
      */
-    inline static factory_fptr_t get_factory(DatatypeIDView datatype_id) noexcept {
-        auto const res = find_map_entry(datatype_id, [](auto const &entry) noexcept {
-            return entry.factory_fptr;
-        });
-
-        return res.has_value() ? *res : nullptr;
-    }
+    [[nodiscard]] static factory_fptr_t get_factory(DatatypeIDView datatype_id) noexcept;
 
     /**
      * Get a to_string function for a datatype. The factory_fptr_t can be used like `std::string str_repr = to_string_fptr(any_typed_arg)`.
      * @param datatype_id datatype id for the corresponding datatype
      * @return function pointer or nullptr
      */
-    inline static to_string_fptr_t get_to_string(DatatypeIDView datatype_id) noexcept {
-        auto const res = find_map_entry(datatype_id, [](auto const &entry) noexcept {
-            return entry.to_string_fptr;
-        });
-
-        return res.has_value() ? *res : nullptr;
-    }
+    [[nodiscard]] static to_string_fptr_t get_to_string(DatatypeIDView datatype_id) noexcept;
 
     /**
      * Try to get the numerical ops function table for a datatype.
@@ -305,23 +235,7 @@ public:
      * @param datatype_id datatype id for the corresponding datatype
      * @return if available a pointer to a structure containing function pointers for all numeric ops; otherwise a null pointer
      */
-    inline static NumericOps const *get_numerical_ops(DatatypeIDView const datatype_id) noexcept {
-        auto const res = find_map_entry(datatype_id, [](auto const &entry) noexcept {
-            return &entry.numeric_ops;
-        });
-
-        // res is nullopt if no datatype matching given datatype_iri was found
-        if (res.has_value()) {
-            // contained ptr cannot be nullptr as by return in lambda for find_map_entry above
-            // optional behind contained ptr can be nullopt if type is not numeric
-            if (auto const ops_ptr = res.value(); ops_ptr->has_value()) {
-                return &ops_ptr->value();
-            }
-        }
-
-        // no datatype found or not numeric
-        return nullptr;
-    }
+    [[nodiscard]] static NumericOps const *get_numerical_ops(DatatypeIDView datatype_id) noexcept;
 
     /**
      * Try to get the conversion function that converts a value of a datatype to it's effective boolean value.
@@ -330,13 +244,7 @@ public:
      * @param datatype_id datatype id for the corresponding datatype
      * @return if available a the conversion function
      */
-    inline static ebv_fptr_t get_ebv(DatatypeIDView const datatype_id) noexcept {
-        auto const res = find_map_entry(datatype_id, [](auto const &entry) noexcept {
-            return entry.ebv_fptr;
-        });
-
-        return res.has_value() ? *res : nullptr;
-    }
+    [[nodiscard]] static ebv_fptr_t get_ebv(DatatypeIDView datatype_id) noexcept;
 
     /**
      * Try to get the comparison function for a datatype.
@@ -345,37 +253,24 @@ public:
      * @param datatype_id datatype id for the corresponding datatype
      * @return if available the comparison function
      */
-    inline static compare_fptr_t get_compare(DatatypeIDView const datatype_id) noexcept {
-        auto const res = find_map_entry(datatype_id, [](auto const &entry) noexcept {
-            return entry.compare_fptr;
-        });
+    [[nodiscard]] static compare_fptr_t get_compare(DatatypeIDView datatype_id) noexcept;
 
-        return res.has_value() ? *res : nullptr;
-    }
-
-    inline static InliningOps const *get_inlining_ops(DatatypeIDView const datatype_id) noexcept {
-        auto const res = find_map_entry(datatype_id, [](auto const &entry) noexcept {
-            return &entry.inlining_ops;
-        });
-
-        // res is nullopt if no datatype matching given datatype_iri was found
-        if (res.has_value()) {
-            // contained ptr cannot be nullptr as by return in lambda for find_map_entry above
-            // optional behind contained ptr can be nullopt if type is not inlineable
-            if (auto const ops_ptr = res.value(); ops_ptr->has_value()) {
-                return &ops_ptr->value();
-            }
-        }
-
-        // no datatype found or not inlineable
-        return nullptr;
-    }
+    /**
+     * Try to get the inlining ops structure for a given datatype.
+     *
+     * @param datatype_id datatype id for the corresponding datatype
+     * @return A pointer to the inlining ops of the given datatype if it exists and fullfills Inlineable, else nullptr
+     */
+    [[nodiscard]] static InliningOps const *get_inlining_ops(DatatypeIDView datatype_id) noexcept;
 
     /**
      * Tries to find a conversion to a common type in the conversion tables lhs_conv and rhs_conv.
      * @return the found conversion if there is a viable one
      */
-    static std::optional<DatatypeConverter> get_common_type_conversion(RuntimeConversionTable const &lhs_conv, RuntimeConversionTable const &rhs_conv, size_t lhs_init_soff = 0, size_t rhs_init_soff = 0) noexcept;
+    [[nodiscard]] static std::optional<DatatypeConverter> get_common_type_conversion(RuntimeConversionTable const &lhs_conv,
+                                                                                     RuntimeConversionTable const &rhs_conv,
+                                                                                     size_t lhs_init_soff = 0,
+                                                                                     size_t rhs_init_soff = 0) noexcept;
 
     /**
      * Tries to find a conversion to a common type in the context of numeric operations.
@@ -385,15 +280,7 @@ public:
      * @note must be called with datatypes that are numeric
      * @return A conversion to a common type that is also a numeric-impl if there is a viable one
      */
-    inline static std::optional<DatatypeConverter> get_common_numeric_op_type_conversion(DatatypeEntry const &lhs_entry, DatatypeEntry const &rhs_entry) noexcept {
-        assert(lhs_entry.numeric_ops.has_value());
-        assert(rhs_entry.numeric_ops.has_value());
-
-        size_t const lhs_init_soff = lhs_entry.numeric_ops->is_stub() ? lhs_entry.numeric_ops->get_stub().start_s_off : 0;
-        size_t const rhs_init_soff = rhs_entry.numeric_ops->is_stub() ? rhs_entry.numeric_ops->get_stub().start_s_off : 0;
-
-        return get_common_type_conversion(lhs_entry.conversion_table, rhs_entry.conversion_table, lhs_init_soff, rhs_init_soff);
-    }
+    [[nodiscard]] static std::optional<DatatypeConverter> get_common_numeric_op_type_conversion(DatatypeEntry const &lhs_entry, DatatypeEntry const &rhs_entry) noexcept;
 
     /**
      * Returns the conversion that turns a value for a numeric-stub DatatypeEntry into
@@ -402,12 +289,7 @@ public:
      * @note must be called with a datatype that is stub-numeric
      * @return the conversion to the corresponding impl-type
      */
-    inline static RuntimeConversionEntry const &get_numeric_op_impl_conversion(DatatypeEntry const &entry) noexcept {
-        assert(entry.numeric_ops.has_value());
-        assert(entry.numeric_ops->is_stub());
-
-        return entry.conversion_table.conversion_at_index(entry.numeric_ops->get_stub().start_s_off, 0);
-    }
+    [[nodiscard]] static RuntimeConversionEntry const &get_numeric_op_impl_conversion(DatatypeEntry const &entry) noexcept;
 
     /**
      * Tries to find a conversion to convert lhs_type_id and rhs_type_id into a
@@ -416,27 +298,37 @@ public:
      * @return nullopt if any of lhs_type_id or rhs_type_id does not have a datatype registered
      * or there is no viable conversion, else the found conversion
      */
-    inline static std::optional<DatatypeConverter> get_common_type_conversion(DatatypeIDView lhs_type_id, DatatypeIDView rhs_type_id) noexcept {
-        auto const lhs_entry = get_entry(lhs_type_id);
-        if (lhs_entry == nullptr) {
-            return std::nullopt;
-        }
-
-        auto const &rhs_entry = get_entry(rhs_type_id);
-        if (rhs_entry == nullptr) {
-            return std::nullopt;
-        }
-
-        return get_common_type_conversion(lhs_entry->conversion_table, rhs_entry->conversion_table);
-    }
+    [[nodiscard]] static std::optional<DatatypeConverter> get_common_type_conversion(DatatypeIDView lhs_type_id, DatatypeIDView rhs_type_id) noexcept;
 };
 
+template<typename Map>
+    requires std::invocable<Map, DatatypeRegistry::DatatypeEntry const &>
+std::optional<std::invoke_result_t<Map, DatatypeRegistry::DatatypeEntry const &>> DatatypeRegistry::find_map_entry(DatatypeIDView const datatype_id, Map f) noexcept(std::is_nothrow_invocable_v<Map, DatatypeEntry const &>) {
 
-/**
- * To register a datatype, at least a LiteralDatatypeImpl instantiation must be provided.
- * If LiteralDatatype_t::cpp_type does not overload `operator<<`, the to_string(const datatype_t &value) must be specialized.
- * @tparam LiteralDatatype_t datatype that is being registered
- */
+    using ret_type = std::optional<std::invoke_result_t<Map, DatatypeEntry const &>>;
+    auto const &registry = registered_datatypes();
+
+    return visit(DatatypeIDVisitor{
+                         [&registry, f](LiteralType const fixed_id) -> ret_type {
+                             auto const id_as_index = static_cast<size_t>(fixed_id.to_underlying()) - 1;  // ids from 1 to n stored in places 0 to n-1
+                             assert(id_as_index < dynamic_datatype_offset);
+
+                             return f(registry[id_as_index]);
+                         },
+                         [&registry, f](std::string_view const other_iri) -> ret_type {
+                             auto found = std::lower_bound(registry.begin() + dynamic_datatype_offset, registry.end(),
+                                                           DatatypeEntry::for_search(other_iri),
+                                                           [](const auto &left, const auto &right) { return left.datatype_iri < right.datatype_iri; });
+
+                             if (found != registry.end() and found->datatype_iri == other_iri) {
+                                 return f(*found);
+                             } else {
+                                 return std::nullopt;
+                             }
+                         }},
+                 datatype_id);
+}
+
 template<datatypes::LiteralDatatype LiteralDatatype_t>
 inline void DatatypeRegistry::add() noexcept {
     using conversion_table_t = decltype(make_conversion_table_for<LiteralDatatype_t>());
@@ -543,7 +435,7 @@ template<typename T>
 }  // namespace detail
 
 template<datatypes::NumericImplLiteralDatatype LiteralDatatype_t>
-inline DatatypeRegistry::NumericOpsImpl DatatypeRegistry::make_numeric_ops_impl() noexcept {
+DatatypeRegistry::NumericOpsImpl DatatypeRegistry::make_numeric_ops_impl() noexcept {
     return NumericOpsImpl{
             // 0
             .zero_value_fptr = []() noexcept -> std::any {
@@ -617,83 +509,6 @@ DatatypeRegistry::InliningOps DatatypeRegistry::make_inlining_ops() noexcept {
         .from_inlined_fptr = [](uint64_t inlined_value) noexcept -> std::any {
             return LiteralDatatype_t::from_inlined(inlined_value);
         }};
-}
-
-inline std::optional<DatatypeRegistry::DatatypeConverter> DatatypeRegistry::get_common_type_conversion(
-        RuntimeConversionTable const &lhs_conv,
-        RuntimeConversionTable const &rhs_conv,
-        size_t const lhs_init_soff,
-        size_t const rhs_init_soff) noexcept {
-
-    auto const find_conv_impl = [](RuntimeConversionTable const &lesser, RuntimeConversionTable const &greater,
-                                   size_t const lesser_init_soff, size_t const greater_init_soff) noexcept -> std::optional<DatatypeConverter> {
-        auto const lesser_s_rank = lesser.subtype_rank() - lesser_init_soff;
-        auto const greater_s_rank = greater.subtype_rank() - greater_init_soff;
-
-        // lesser should be the conversion table of the type with lower subtype rank
-        assert(lesser_s_rank <= greater_s_rank);
-
-        // calculate initial subtype offsets to equalize subtype rank
-        size_t lesser_s_off = lesser_init_soff;
-        size_t greater_s_off = greater_init_soff + greater_s_rank - lesser_s_rank;
-
-        while (lesser_s_off < lesser.subtype_rank() && greater_s_off < greater.subtype_rank()) {
-            auto const lesser_p_rank = lesser.promotion_rank_at_level(lesser_s_off);
-            auto const greater_p_rank = greater.promotion_rank_at_level(greater_s_off);
-
-            if (lesser_p_rank == greater_p_rank) {
-                // subtype rank and promotion rank equal
-                // => potential for correct conversion
-
-                RuntimeConversionEntry const &lconv = lesser.conversion_at_index(lesser_s_off, 0);
-                RuntimeConversionEntry const &gconv = greater.conversion_at_index(greater_s_off, 0);
-
-                if (lconv.target_type_id == gconv.target_type_id) {
-                    // correct conversion found
-                    return DatatypeConverter::from_individuals(lconv, gconv);
-                }
-            } else {
-                // subtype rank equal
-                // promotion rank not yet equal
-
-                // calculate promotion offsets to equalize promotion rank
-                auto const [lesser_p_off, greater_p_off] = lesser_p_rank < greater_p_rank
-                                                                   ? std::make_pair(0ul, greater_p_rank - lesser_p_rank)
-                                                                   : std::make_pair(lesser_p_rank - greater_p_rank, 0ul);
-
-                if (lesser_p_off < lesser_p_rank && greater_p_off < greater_p_rank) {
-                    RuntimeConversionEntry const &lconv = lesser.conversion_at_index(lesser_s_off, lesser_p_off);
-                    RuntimeConversionEntry const &gconv = greater.conversion_at_index(greater_s_off, greater_p_off);
-
-                    if (lconv.target_type_id == gconv.target_type_id) {
-                        // correct conversion found
-                        return DatatypeConverter::from_individuals(lconv, gconv);
-                    }
-                }
-            }
-
-            lesser_s_off += 1;
-            greater_s_off += 1;
-        }
-
-        // no conversion available
-        return std::nullopt;
-    };
-
-    // call find_conv_impl with entries in correct order (lesser s rank, greater s rank)
-    if (lhs_conv.subtype_rank() - lhs_init_soff < rhs_conv.subtype_rank() - rhs_init_soff) {
-        return find_conv_impl(lhs_conv, rhs_conv, lhs_init_soff, rhs_init_soff);
-    } else {
-        auto res = find_conv_impl(rhs_conv, lhs_conv, rhs_init_soff, lhs_init_soff);
-
-        if (res.has_value()) {
-            // swap functions to reverse the ordering change
-            std::swap(res->convert_lhs, res->convert_rhs);
-            std::swap(res->inverted_convert_lhs, res->inverted_convert_rhs);
-        }
-
-        return res;
-    }
 }
 
 }  // namespace rdf4cpp::rdf::datatypes::registry
